Avoid permission issues when renewing under cron
Dehydrated is kicked off by the root user's crontab and so with a current working directory of /root. It then su-exec's to the UID/GID provided in the environment. Without specifying a config-dir lexicon uses the current working directory. As not-root it doesn't have permissions to look for YAML files in /root.

Specify an accessible directory to use as the config directory. We're not expecting to find YAML files there, we just need it to be readable by UID/GID. By using `/data` it is possible for us to provide YAML files if we so desire